Move riffle_shuffle insto its own file
