+++ packages/std/src/testing/mod.rs
@@ -5,6 +5,7 @@
 
 mod assertions;
 mod mock;
+mod shuffle;
 
 pub use assertions::assert_approx_eq_impl;
 
@@ -12,7 +13,7 @@ pub use assertions::assert_approx_eq_impl;
 pub use mock::StakingQuerier;
 pub use mock::{
     digit_sum, mock_dependencies, mock_dependencies_with_balance, mock_dependencies_with_balances,
-    mock_env, mock_info, mock_wasmd_attr, riffle_shuffle, BankQuerier, MockApi, MockQuerier,
+    mock_env, mock_info, mock_wasmd_attr, BankQuerier, MockApi, MockQuerier,
     MockQuerierCustomHandlerResult, MockStorage, MOCK_CONTRACT_ADDR,
 };
 #[cfg(feature = "stargate")]
@@ -21,3 +22,4 @@ pub use mock::{
     mock_ibc_channel_connect_ack, mock_ibc_channel_connect_confirm, mock_ibc_channel_open_init,
     mock_ibc_channel_open_try, mock_ibc_packet_ack, mock_ibc_packet_recv, mock_ibc_packet_timeout,
 };
+pub use shuffle::riffle_shuffle;
